Replace lsb_release call with parsing /etc/os-release
The lsb project is in an abonded state and some distros have already dropped support for the `lsb_release` command (notably RHEL 9 and clones). It can be suspected that more distros (including Ubuntu) might drop support for it. Therefore: Replace lsb_release call with parsing /etc/os-release which is provided with all major distros.